+++ spynnaker/pyNN/models/utility_models/delays/delay_extension_vertex.py
@@ -64,7 +64,7 @@ class DelayExtensionVertex(
             source_vertex, constraints=None, label="DelayExtension"):
         """
         :param int n_neurons: the number of neurons
-        :param int delay_per_stage: the delay per stage
+        :param float delay_per_stage: the delay per stage
         :param int max_delay_to_support: the max delay this will cover
         :param ~pacman.model.graphs.application.ApplicationVertex \
                 source_vertex:
